backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
alembic
sqlalchemy-utils
passlib[bcrypt]
pydantic[email]
python-jose[cryptography]
python-multipart
ultralytics
Pillow
google-cloud-vision
transformers[torch]
Shapely